refactor(library/modifiersByType): moves "/error/Contextualized.ts" to "/error/Contextualized/definition.ts"
